+++ KmpIO/src/androidUnitTest/kotlin/com/oldguy/common/io/TimeTests.kt
@@ -1,6 +1,7 @@
 package com.oldguy.common.io
 
 import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.number
 import org.junit.Test
 import java.util.*
 import java.util.zip.ZipEntry
@@ -25,18 +26,18 @@ class TimeTests {
         val cal = GregorianCalendar(y, jm.toInt(), d.toInt(), h, m, s)
         val dt = cal.time
         val e = ZipEntry("x")
-        e.setTime(dt.time)
+        e.time = dt.time
         val cal2 = GregorianCalendar.getInstance().apply { time = Date(e.time) }
         ZipTime(LocalDateTime(2017, 6, 23, 18, 46, s)).apply {
             assertEquals(cal.get(Calendar.YEAR), zipTime.year)
-            assertEquals(cal.get(Calendar.MONTH) + 1, zipTime.monthNumber)
-            assertEquals(cal.get(Calendar.DAY_OF_MONTH), zipTime.dayOfMonth)
+            assertEquals(cal.get(Calendar.MONTH) + 1, zipTime.month.number)
+            assertEquals(cal.get(Calendar.DAY_OF_MONTH), zipTime.day)
             assertEquals(cal.get(Calendar.HOUR_OF_DAY), zipTime.hour)
             assertEquals(cal.get(Calendar.MINUTE), zipTime.minute)
             assertEquals(cal.get(Calendar.SECOND), zipTime.second)
             assertEquals(cal2.get(Calendar.YEAR), zipTime.year)
-            assertEquals(cal2.get(Calendar.MONTH) + 1, zipTime.monthNumber)
-            assertEquals(cal2.get(Calendar.DAY_OF_MONTH), zipTime.dayOfMonth)
+            assertEquals(cal2.get(Calendar.MONTH) + 1, zipTime.month.number)
+            assertEquals(cal2.get(Calendar.DAY_OF_MONTH), zipTime.day)
             assertEquals(cal2.get(Calendar.HOUR_OF_DAY), zipTime.hour)
             assertEquals(cal2.get(Calendar.MINUTE), zipTime.minute)
             assertEquals(cal2.get(Calendar.SECOND), zipTime.second)